pyproject.toml
[build-system]
requires = [ "setuptools>=61.0", "wheel",]
build-backend = "setuptools.build_meta"

[project]
name = "huginn"
version = "0.0.1"
description = ""
readme = "README.md"
requires-python = ">=3.8"
dependencies = ["accelerate", "hf_xet", "huggingface_hub", "noexcept", "python-dotenv", "torch", "transformers"]
[[project.authors]]
name = "Nichola Walch"
email = "littler.compression@gmail.com"

[project.license]
file = "LICENSE"

[project.scripts]
aiaio = "aiaio.cli:main"
